Listado de Productos FUNCIONAL y archivo PROCEDIMIENTOS

+++ ProyectoRefri/Models/ProductoModel.cs
@@ -6,16 +6,22 @@ namespace ProyectoRefri.Models
     {
         [Required][Display (Name ="Id")]
         public string idProducto{get;set;}
+
         [Required][Display(Name = "Nombre")]
         public string nombre { get; set; }
+
         [Required][Display(Name = "Precio")]
-        public float precio { get; set; }
+        public double precio { get; set; }
+
         [Required][Display(Name = "Stock")]
         public Int32 stock { get; set; }
+
         [Display(Name = "Descripcion")]
         public string descripcion { get; set; }
+
         [Required][Display(Name = "Imagen")]
         public string imagen { get; set; }
+
         [Display(Name = "Estado")]
         public int estado { get; set; }
